Added changelog, removed duplicate residence variable, rolled major version number
