+++ Procfile
@@ -1,1 +1,1 @@
-web: SERVER_NAME=beststart.herokuapp.com openfisca serve --country-package openfisca_aotearoa --port $PORT --bind 0.0.0.0:$PORT
+web: SERVER_NAME=openfisca-aotearoa.herokuapp.com openfisca serve --country-package openfisca_aotearoa --port $PORT --bind 0.0.0.0:$PORT
